Add routine to compute stochastic error probability
In a stochastic error model (the model with best rigorous
threshold bounds in AGP06), the error map must be written as a
mixture of the identity channel and some other CPTP map.

The diamond
norm does not guarantee this, so instead I am adding a simple routine
that, given some channel, it will find the smallest probability
of error that guarantees that the error map can be written as the mixture of
the identity and some arbitrary CPTP map.

This is aprticularly useful for error models that are not just
Pauli errors, e.g., that include small unitary errors as well as
some decoherence.

Note that this metric is very pessimistic. Any purely unitary error away from
the identity leads to a error probability of 1.